Refactor Missions and MissionsList components
